requirements.txt
accelerate==0.18.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bitsandbytes @ https://github.com/jllllll/bitsandbytes-windows-webui/raw/main/bitsandbytes-0.38.1-py3-none-any.whl#sha256=b8110e3360d80f1694f59339a680a37a4003c72edb59002528333e06295623b5
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.7
datasets==2.10.1
decorator==5.1.1
dill==0.3.6
diskcache==5.6.1
dnspython==2.3.0
docarray==0.21.0
docker==6.1.2
docopt==0.6.2
duckduckgo-search==2.9.1
ecdsa==0.18.0
elastic-transport==8.4.0
elasticsearch==8.7.0
email-validator==2.0.0.post2
entrypoints==0.4
fairscale==0.4.13
fastapi==0.95.1
ffmpy==0.3.0
filelock==3.12.0
fire==0.5.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
hiq-python==1.1.12
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
InstructorEmbedding==1.0.0
itsdangerous==2.1.2
jina-hubble-sdk==0.36.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.163
linkify-it-py==2.0.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
openapi-schema-pydantic==1.2.4
orjson==3.8.11
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
Pillow==9.5.0
protobuf==3.20.3
psutil==5.9.5
py-itree==0.0.19
pyarrow==12.0.0
pyasn1==0.5.0
pydantic==1.10.7
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.15.1
pyllama==0.0.9
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0
quant-cuda @ https://github.com/jllllll/GPTQ-for-LLaMa-Wheels/raw/main/quant_cuda-0.0.0-cp310-cp310-win_amd64.whl#sha256=8b8b70ecaddcdcda28c4697f7d31dc6b8f3331f3e1d9e6a6669c3e467b5fc979
regex==2023.5.5
requests==2.29.0
responses==0.18.0
rich==13.3.5
rouge==1.0.1
rsa==4.9
safetensors==0.3.0
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.12
starlette==0.26.1
sympy==1.11.1
tenacity==8.2.2
termcolor==2.3.0
threadpoolctl==3.1.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tornado==6.3.1
tqdm==4.65.0
transformers==4.28.0
types-requests==2.30.0.0
types-urllib3==1.26.25.13
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uc-micro-py==1.0.2
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.22.0
validators==0.20.0
watchdog==3.0.0
watchfiles==0.19.0
websocket-client==1.5.1
websockets==11.0.2
wikipedia==1.4.0
xxhash==3.2.0
yarg==0.1.9
yarl==1.9.2
zipp==3.15.0
yaspin==2.3.0
